src/modules/cars/useCases/CreateCarSpecification/CreateCarSpecificationUseCase.spec.ts
import { SpecificationsRepositoryInMemory } from "@modules/cars/repositories/in-memory/SpecificationsRepositoryInMemory";
import { AppError } from "@shared/errors/AppError";
import { CarsRepositoryInMemory } from "../../repositories/in-memory/CarsRepositoryInMemory";
import { CreateCarSpecificationUseCase } from "./CreateCarSpecificationUseCase";

let createCarSpecificationUseCase: CreateCarSpecificationUseCase;
let carsRepositoryInMemory: CarsRepositoryInMemory;
let specificationsRepositoryInMemory: SpecificationsRepositoryInMemory;

describe("Create Car Specifications", () => {
  beforeAll(() => {
    carsRepositoryInMemory = new CarsRepositoryInMemory();
    specificationsRepositoryInMemory = new SpecificationsRepositoryInMemory();
    ("");
    createCarSpecificationUseCase = new CreateCarSpecificationUseCase(
      carsRepositoryInMemory,
      specificationsRepositoryInMemory,
    );
  });

  it("Should not be able to create a new car specification to a nonexistent car ", async () => {
    expect(async () => {
      const car_id = "123";
      const specifications_id = ["123456"];

      await createCarSpecificationUseCase.execute({
        car_id,
        specifications_id,
      });
    }).rejects.toBeInstanceOf(AppError);
  });

  it("Should not be able to create a new car specification", async () => {
    const car = {
      name: "Car Test one",
      description: "Car description Test",
      daily_rate: 1,
      license_plate: "123123",
      fine_amount: 10,
      brand: "Car Brand Test",
      category_id: "category_id",
    };

    const specification = await specificationsRepositoryInMemory.create({
      name: "specification",
      description: "Specification description",
    });

    const createdCar = await carsRepositoryInMemory.create(car);

    const specificationCars = await createCarSpecificationUseCase.execute({
      car_id: createdCar.id,
      specifications_id: [specification.id],
    });

    expect(specificationCars).toHaveProperty("specifications");
    expect(specificationCars.specifications.length).toBe(1);
  });
});
